Updated CHANGELOG.md and improved coverage.
Also removed debug prints.
