+++ tests/conftest.py
@@ -95,18 +95,6 @@ def make_line_collection() -> LineCollectionMaker:
     return _make_line_collection
 
 
-def test_make_line_collection(make_line_collection):
-    lc = make_line_collection(with_metadata=False)
-    assert lc.metadata == {}
-    assert len(lc) > 0
-
-    lc = make_line_collection(with_metadata=True)
-    assert vp.METADATA_FIELD_NAME in lc.metadata
-    assert isinstance(lc.metadata[vp.METADATA_FIELD_COLOR], vp.Color)
-    assert isinstance(lc.metadata[vp.METADATA_FIELD_PEN_WIDTH], float)
-    assert len(lc) > 0
-
-
 class DocumentMaker(Protocol):
     def __call__(self, layer_count: int = ...) -> vp.Document:
         ...
@@ -122,17 +110,6 @@ def make_document(make_line_collection) -> DocumentMaker:
         return doc
 
     return _make_document
-
-
-def test_make_document(make_document):
-    doc = make_document(layer_count=3)
-    assert len(doc.layers) == 3
-    assert doc.page_size is not None
-    for lc in doc.layers.values():
-        assert vp.METADATA_FIELD_NAME in lc.metadata
-        assert isinstance(lc.metadata[vp.METADATA_FIELD_COLOR], vp.Color)
-        assert isinstance(lc.metadata[vp.METADATA_FIELD_PEN_WIDTH], float)
-        assert len(lc) > 0
 
 
 # IMAGE COMPARE SUPPORT
