+++ vpype_cli/blocks.py
@@ -175,11 +175,9 @@ def forfile(state: State, processors: Iterable[ProcessorType], files: str) -> No
                 write "optimized/%basename(_path)%" end
     """
 
-    print(files)
     file_list = sorted(
         glob.glob(os.path.expandvars(os.path.expanduser(files))), key=os.path.basename
     )
-    print(file_list)
     orig_document = state.document
 
     for i, file in enumerate(file_list):
